+++ src/main/java/com/programmers/vouchermanagement/voucher/domain/Voucher.java
@@ -2,11 +2,32 @@ package com.programmers.vouchermanagement.voucher.domain;
 
 import java.util.UUID;
 
-public interface Voucher {
+public class Voucher {
 
-    UUID getVoucherId();
+    private final UUID voucherId;
+    private final Long discount;
+    private final VoucherType voucherType;
+    private final VoucherPolicy voucherPolicy;
 
-    VoucherType getVoucherType();
+    public Voucher(UUID voucherId, Long discount, VoucherType voucherType, VoucherPolicy voucherPolicy) {
 
-    Long getDiscount();
+        voucherPolicy.validateDiscount(discount);
+
+        this.voucherId = voucherId;
+        this.discount = discount;
+        this.voucherType = voucherType;
+        this.voucherPolicy = voucherPolicy;
+    }
+
+    public UUID getVoucherId() {
+        return voucherId;
+    }
+
+    public Long getDiscount() {
+        return discount;
+    }
+
+    public VoucherType getVoucherType() {
+        return voucherType;
+    }
 }
