+++ src/main/java/com/programmers/vouchermanagement/voucher/domain/VoucherPolicy.java
@@ -1,0 +1,6 @@
+package com.programmers.vouchermanagement.voucher.domain;
+
+public interface VoucherPolicy {
+
+    void validateDiscount(Long discount);
+}
